+++ NEWS.md
@@ -1,7 +1,8 @@
 # Changes in release 4.4.0
 
 - If 'verbose' is 'TRUE' and all targets are already up to date (nothing to build), then `make()` and `outdated()` print "All targets are already up to date" to the console.
-- We added powerful parallelism options through the `future` and `future.batchtools` packages. The vignettes walk through the new functionality.
+- New support for timeouts and retries when it comes to building targets.
+- Failed targets are now recorded during the build process. You can see them in `plot_graph()` and `progress()`. Also see the new `failed()` function, which is similar to `in_progress()`.
 
 # 2017-10-17